feat: add login verification in User model #50
Added 'getUserByLoginPass' method to the User model. This method checks if there is a user in the database that matches a given email and hashed password. If a match is found, it returns the user's information (except the password) as an object. If no match is found, it throws an exception.